% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_to_num.R
\name{e_to_num}
\alias{e_to_num}
\title{Remove space, change decimal sign and make numeric}
\usage{
e_to_num(x)
}
\arguments{
\item{x}{A vector to tidy and make numeric}
}
\value{
A numeric value
}
\description{
Remove blank space, swap "," to "." and convert to nummeric
}

