% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfix-pkg.R
\docType{package}
\name{qfix}
\alias{qfix}
\alias{qfix-package}
\title{qfix: A collection of misc quick fix functions}
\description{
A collection of probably poor documented and possibly unstable
funcitons that's might be useful anyway to ME. Use at own risk, functions might
be randomly added, deprecated or broken from time to time.
A lot of (my) common used packages is listed as Imports: in DESCRIPTION to
to ensure they're installed.
}
\author{
Stefan Jansson \email{janzzon@gmail.com}
}

